feat(worker): add journal sync worker; minor fixes

+++ src/infrastructure/adapters/outbound/database/migrations/20230524170858_create-table-journal.ts
@@ -6,8 +6,8 @@ export async function up(knex: Knex): Promise<void> {
   return knex.schema.createTable(tableName, (table) => {
     table.increments('id', { primaryKey: true });
 
-    table.integer('chain_id').notNullable();
-    table.integer('block_number').notNullable();
+    table.bigInteger('chain_id').notNullable();
+    table.bigInteger('block_number').notNullable();
     table.string('transaction_hash').notNullable();
 
     table.decimal('amount', 20, 8).notNullable();
@@ -17,12 +17,12 @@ export async function up(knex: Knex): Promise<void> {
     table.string('account').notNullable();
     table.string('account_group').notNullable();
     table.string('iso_code').notNullable();
-    table.integer('log_index').notNullable();
+    table.bigInteger('log_index').notNullable();
     table.datetime('entry_date').notNullable();
 
-    table.integer('gas_used').nullable();
-    table.integer('cumulative_gas_used').nullable();
-    table.integer('effective_gas_price').nullable();
+    table.bigInteger('gas_used').nullable();
+    table.bigInteger('cumulative_gas_used').nullable();
+    table.bigInteger('effective_gas_price').nullable();
 
     table.datetime('created_at').defaultTo(knex.fn.now());
   });
